Work around broken MPI installations
CMake's FindMPI.cmake is not robust enough to cope with broken MPI
installations that might be found and no upstream fix is available
yet. This is a problem when the optional dependency on MPI for gmxapi
causes configuration to fail even when the user does not intend to use
MPI at all. Now CMake suggests a workaround if the user does encounter
such problems.

Fixes #4699